+++ TODO.R
@@ -1,3 +1,2 @@
-# TODO [todo] todo could guess where it should be written. Maybe om top of an R file in the project?
 # TODO [screenshot] make the behaviour different when vignettes vs articles: vignettes should place it in man/figures, while articles could put it in vignettes/articles file.
 # TODO [screenshot] RStudio addin to insert the code directly in the qmd doc.